@@ -32,53 +32,53 @@
   },
   "homepage": "https://github.com/react-bootstrap/react-router-bootstrap",
   "peerDependencies": {
-    "react": ">=0.13.0",
-    "react-router": ">=1.0.0-rc3"
+    "react": ">=0.14.0",
+    "react-router": ">=2.0.0-rc2"
   },
   "devDependencies": {
-    "babel": "^5.8.23",
-    "babel-core": "^5.8.25",
-    "babel-eslint": "^4.1.3",
-    "babel-loader": "^5.3.2",
-    "bootstrap": "^3.3.5",
+    "babel": "5.x",
+    "babel-core": "5.x",
+    "babel-eslint": "^4.1.6",
+    "babel-loader": "5.x",
+    "bootstrap": "^3.3.6",
     "colors": "^1.1.2",
-    "css-loader": "^0.19.0",
-    "es5-shim": "^4.1.14",
-    "eslint": "^1.7.1",
-    "eslint-config-airbnb": "^0.1.0",
-    "eslint-plugin-babel": "^2.1.1",
-    "eslint-plugin-mocha": "^1.0.0",
-    "eslint-plugin-react": "^3.5.1",
-    "file-loader": "^0.8.4",
-    "history": "^1.12.5",
-    "html-webpack-plugin": "^1.6.2",
-    "karma": "^0.13.11",
-    "karma-cli": "^0.1.1",
-    "karma-mocha": "^0.2.0",
-    "karma-mocha-reporter": "^1.1.1",
-    "karma-phantomjs-launcher": "^0.2.1",
+    "css-loader": "^0.23.1",
+    "es5-shim": "^4.4.1",
+    "eslint": "^1.10.3",
+    "eslint-config-airbnb": "^2.1.1",
+    "eslint-plugin-babel": "^3.0.0",
+    "eslint-plugin-mocha": "^1.1.0",
+    "eslint-plugin-react": "^3.13.1",
+    "file-loader": "^0.8.5",
+    "history": "^1.17.0",
+    "html-webpack-plugin": "^1.7.0",
+    "karma": "^0.13.16",
+    "karma-cli": "^0.1.2",
+    "karma-mocha": "^0.2.1",
+    "karma-mocha-reporter": "^1.1.5",
+    "karma-phantomjs-launcher": "^0.2.2",
     "karma-sinon-chai": "^1.1.0",
     "karma-sourcemap-loader": "^0.3.6",
     "karma-webpack": "^1.7.0",
     "less": "^2.5.3",
-    "less-loader": "^2.2.1",
+    "less-loader": "^2.2.2",
     "lodash": "^3.10.1",
-    "mocha": "^2.3.3",
+    "mocha": "^2.3.4",
     "mt-changelog": "^0.6.2",
     "node-libs-browser": "^0.5.3",
-    "phantomjs": "^1.9.18",
-    "react": "^0.14.0",
-    "react-bootstrap": "^0.27.2",
-    "react-dom": "^0.14.0",
-    "react-router": "^1.0.0-rc3",
-    "release-script": "^0.5.4",
-    "rimraf": "^2.4.3",
+    "phantomjs": "^1.9.19",
+    "react": "^0.14.5",
+    "react-bootstrap": "^0.28.1",
+    "react-dom": "^0.14.5",
+    "react-router": "^2.0.0-rc4",
+    "release-script": "^0.5.6",
+    "rimraf": "^2.5.0",
     "shelljs": "^0.5.3",
-    "style-loader": "^0.12.4",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.12.2",
-    "webpack-dev-server": "^1.12.0",
-    "yargs": "^3.27.0"
+    "style-loader": "^0.13.0",
+    "url-loader": "^0.5.7",
+    "webpack": "^1.12.9",
+    "webpack-dev-server": "^1.14.0",
+    "yargs": "^3.31.0"
   },
   "files": [
     "README",
