@@ -1,7 +1,7 @@
 import React from 'react';
 import ButtonToolbar from 'react-bootstrap/lib/ButtonToolbar';
 import Button from 'react-bootstrap/lib/Button';
-import {Link} from 'react-router';
+import { Link } from 'react-router';
 
 import LinkContainer from '../../src/LinkContainer';
 
