@@ -4,7 +4,7 @@ export default {
   },
   module: {
     loaders: [
-      {test: /\.js/, loader: 'babel', exclude: /node_modules/}
+      { test: /\.js/, loader: 'babel', exclude: /node_modules/ }
     ]
   },
   devtool: 'inline-source-map'
